Rename tasnet.py and or_pit.py to model.py for consistency
