+++ padertorch/contrib/examples/source_separation/or_pit/__init__.py
@@ -1,1 +1,1 @@
-from .or_pit import OneAndRestPIT
+from .model import OneAndRestPIT
