+++ padertorch/contrib/examples/source_separation/tasnet/__init__.py
@@ -1,2 +1,2 @@
 from . import tas_coders
-from .tasnet import TasNet
+from .model import TasNet
